Fix styling issues for Calendar
Introduce a fix such that the calendar page no longer overwrites the side bar's styling when we click to another page
